Refactor feature components into index files and reorganize project structure
- Move components into centralized index files for auth, dashboard, and user features
- Update import statements across multiple files to use new index exports
- Remove individual component files to simplify project structure
- Update test files to reflect new import paths
