+++ components/layout/Header.tsx
@@ -1,7 +1,7 @@
 import { TouchableOpacity, View } from "react-native";
 import { usePathname } from "expo-router";
 
-import { AvatarDisplay } from "@/feature/user/components/AvatarDisplay";
+import { AvatarDisplay } from "@/feature/user/components";
 import { useUserContext } from "@/feature/user/contexts/UserContext";
 import { MenuIcon } from "../icons/MenuIcon";
 import { ThemedText } from "../text/ThemedText";
