+++ feature/auth/screen/Login.tsx
@@ -4,9 +4,7 @@ import { View } from "react-native";
 import { PrimaryButton } from "@/components/button/PrimaryButton";
 import { DefaultLink } from "@/components/link/DefaultLink";
 import { ThemedText } from "@/components/text/ThemedText";
-import { AuthLayout } from "../components/AuthLayout";
-import { EmailInput } from "../components/EmailInput";
-import { PasswordInput } from "../components/PasswordInput";
+import { AuthLayout, EmailInput, PasswordInput } from "../components";
 import { loginWithEmail } from "../service/authService";
 
 export const Login = () => {
