+++ feature/dashboad/components/index.tsx
@@ -1,0 +1,5 @@
+// Stats
+export { StatCard } from "./stats/StatCard";
+
+// Views
+export { TopView } from "./views/TopView";
