+++ feature/user/components/index.tsx
@@ -1,0 +1,6 @@
+// Avatar
+export { AvatarDisplay } from "./avatar/AvatarDisplay";
+export { AvatarPicker } from "./avatar/AvatarPicker";
+
+// Actions
+export { UpdateProfileButton } from "./actions/UpdateProfileButton";
